Move Token to Neon
